New Plasmoid: EmailMessage
A plasmoid to display a single email message, as popup or planar.
Depending on the size, it will show more or less information. The
basics for it are now in place. We'll also use this as the delegate
for the new Akonadi applet (but you'll also be able to drop an email
onto panel or desktop, dashboard ...

svn path=/trunk/playground/base/plasma/applets/plasmobiff/; revision=914697