New name for the Akonadi applet: Lion Mail
LionMail is the Plasma UI for Akonadi. The idea is to show collections
that represent live queries to the Akonadi database. The queries sit
inside extenders, so such a collection can be put on the desktop, in the
panel, etc ...

The applet is only half way there right now, the layout is messed up,
but we did get the Akonadi dataengine fixed, and have basically
established end-to-end connection from data that gets into Akonadi,
queried by the akonadi dataengine, and offered to the Plasma applet as
updates.

For this purpose, we use two applets:

LionMail, the container showing collections as groups of emails in
extenderitems and MailMessage. It's supposed to have some collections
set up by default ("Unread Emails", "Emails with tag /Important/ or
/TODO/", ...). Those collections can be configured through the UI that
Akonadi will deliver for it (much like we show KCM's in a "More..."
button in other applet, such as powerdevil and networkmanager). We can
also think of configuring notification for events in certain folders.

MailMessage is an applet that shows a view on an email inside Akonadi.
The idea is to make the applet resizable. When there's enough space,
or it has been resized, it displays more or less information. We should
be able to use those applets inside LionMail, in listviews for example
(zoomable listviews, even), and on the desktop. That way it'll be easy
to drop an email for example onto your desktop for reference.

Finally, the name is obviously a worshipping reference to Camp KDE on
Jamaica.

CCMAIL:plasma-devel@kde.org
CCMAIL:adam@kde.org

svn path=/trunk/playground/base/plasma/applets/plasmobiff/; revision=914921