A lot of useless futzing with the layouting
I'm trying to get the sizechanges from the extenders propagated up to
the popup, to no avail. Seems like I should just put this aside, wait
for Qt 4.5 and see if that changes anything. The surgery needed to get
QGraphicsLayout to properly resize when one the contents of a cell
changes size looks so icky, and it's so frustrating that I'm close to
just giving up for Qt 4.4 and declare it broken beyond reasonable to
fix. The combination of dynamic layout (emailmessage) inside a dynamic
layout (lionmail's MailExtender) inside an Extender inside a popup seems
too much for either me, Qt, or -- quite likely -- both.

I've not tried 4.5 yet, but I have high hopes for it. (If that doesn't
relieve us poor Plasma hackers, I'll probably just jump off a cliff, or
equivalent.)

svn path=/trunk/playground/base/plasma/applets/lionmail/; revision=917366