Some layouting hacks
The graphicswidget now emits a signal when its geometry changed so we
can resize containers above the widget. It does *something*, but not
work yet.

The goal here is to have the applet grow and shrink as one expands it
with the button. (It does already display and hide various pieces of
information, such as recipient and body when being resized, though.)

svn path=/trunk/playground/base/plasma/applets/lionmail/; revision=920471