Fix initial sizing of the emailwidgets in the collection view
Also give some feedback if the syncjob worked out. In my testing it now
all works reliably: Clicking on the action buttons starts a modification
job to sync the new flags to Akonadi. The dataengine monitors the item
and reacts to the change in Akonadi. The Lion Mail applet catches the
change notification and updates the email widget accordingly.
On top of that, the statistics are updated via another signal.
This also works when emails are edited from another application (try
akonadiconsole for example).

The initial sizing of the emailwidget is now also OK. The Tiny and Small
sizes work rather well. For Medium, it needs some more tweaking as the
header isn't always the same height, this needs to be taken into account.

svn path=/trunk/playground/base/plasma/applets/lionmail/; revision=952496