Offer email message for dropped akonadi: URLs
This is actually pretty cool and possible thanks to the neat URL and mimetype handling
in Plasma::Containment we released with 4.4. We just add a wildcard to an akonadi: URL
and we can properly drag and drop email widgets pointing to akonadi items in Plasma.

I love it when a plan comes together. :-)

svn path=/trunk/playground/base/plasma/applets/lionmail/; revision=1117785