New Emailnotifier Applet
This is so to say Lion Mail 2, based on the EntityTreeModel. I'm slowly
simplifying the Lion Mail codebase, but don't want to slaughter that one
as it might have its own usecases in its current form.

The Emailnotifier applet is pretty crude still and lacking most
features. It's the starting point of what we've discussed during Akademy
in the PIM team.

svn path=/trunk/playground/base/plasma/applets/lionmail/; revision=1150471