Do not delete the config group before saving
This makes the config group re-usable because it doesn't destroy
previous data. I'm not sure why this one needs to be cleaned, though.
Steve, any pointer? I'm re-using the Plasma applet's config() here to
save the model's state into the Plasma config, so it happens per applet.
This is I think quite elegant, thanks to the design of the
modelstatesaver.

CCMAIL:steveire@gmail.com

svn path=/trunk/playground/base/plasma/applets/lionmail/; revision=1164523