Use KSelectionProxyModel for the config
This is still broken, it doesn't show any connections in the config's
treeview. I thought a good first step would be getting rid of the copy
(which I'm doing with this commit), but can't find the reason for the
treeview not being populated. The email item etms work fine, I'm not
sure wether it's the config's ETM, the stack of proxy, check and filter
models on top of that, or something else that I'm doing wrong ... :/

Steve, can you have a look at what I'm doing wrong?

CCMAIL:steveire@gmail.com

svn path=/trunk/playground/base/plasma/applets/lionmail/; revision=1178822