Make deleting emails work, add mark as spam & todo
Two things are interesting in this commit:
1) We use the deleted flag for deleting emails, let's see how clients
handle this. I *think* this solution is quite elegant, but we might want
to move items into the trashfolder (which?!) instead. SpecialCollection
might help here. Or maybe we want to _really_ delete items?

2) The context menu offers some additional options to mark emails, spam
or task. These options would clutter the UI too much, but are nice to
have available for power users (those are more likely to find them there)

svn path=/trunk/playground/base/plasma/applets/lionmail/; revision=1200118